requirements.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.1
dill==0.3.2
dm-tree==0.1.5
docker-pycreds==0.4.0
frozendict==1.2
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.9
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
importlib-resources==3.0.0
jax==0.2.4
jaxlib==0.1.56
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3
neural-tangents==0.3.4
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.3.0
pathtools==0.1.2
promise==2.3
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
sentry-sdk==0.18.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-datasets==4.0.1
tensorflow-estimator==2.2.0
tensorflow-metadata==0.24.0
termcolor==1.1.0
tqdm==4.50.2
urllib3==1.25.10
watchdog==0.10.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0